Factor load generation functionality out of Simulation.
