+++ src/ledger/LedgerPerformanceTests.cpp
@@ -84,7 +84,7 @@ class LedgerPerformanceTests : public Simulation
         uint64_t amount = static_cast<uint64_t>(
             rand_fraction() *
             min(static_cast<uint64_t>(1000),
-                (mAccounts[iFrom]->mBalance - getMinBalance()) / 3));
+                (mAccounts[iFrom]->mBalance - mMinBalance) / 3));
         txs.push_back(make_optional<TxInfo>(
             createTransferTransaction(iFrom, iTo, amount)));
